requirements.txt
Flask == 2.0.2
Flask-Restful == 0.3.9
pyjwt==2.3.0
email_validator==1.1.3
python-dotenv==0.20.0
firebase-admin==5.2.0
twilio
cryptocode
password-validator